feat: 코드 리팩토링 및 전체 레이아웃 수정
